Add WidgetApi frontstage.
