+++ test-apps/appui-test-app/appui-test-providers/src/appui-test-providers.ts
@@ -3,13 +3,13 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 export * from "./tools/ContentLayoutTools";
+export * from "./tools/RegisterUiProviderTool";
 
 export * from "./ui/components/LanguageSelect";
 
 export * from "./ui/dialogs/SampleModalDialog";
 
 export * from "./ui/providers/AbstractUiItemsProvider";
-export * from "./ui/providers/CustomStageUiItemsProvider";
 export * from "./ui/providers/FloatingWidgetsUiItemsProvider";
 export * from "./ui/providers/InspectUiItemInfoToolProvider";
 export * from "./ui/providers/MessageUiItemsProvider";
@@ -23,7 +23,11 @@ export * from "./ui/frontstages/CustomFrontstageProvider";
 export * from "./ui/frontstages/PopoutWindowsFrontstage";
 export * from "./ui/frontstages/registerCustomFrontstage";
 export * from "./ui/frontstages/SynchronizedFloatingViewport";
-export * from "./ui/frontstages/WidgetApiStage";
+
+export * from "./ui/widgets/ITwinUIv2Widget";
+export * from "./ui/widgets/LayoutWidget";
+export * from "./ui/widgets/LogLifecycleWidget";
+export * from "./ui/widgets/UseWidgetHookWidget";
 
 export * from "./ui/ViewportContent";
