--- test-apps/appui-test-app/appui-test-providers/src/ui/providers/CustomStageUiItemsProvider.tsx
@@ -1,40 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-
-import {
-  BackstageItem,
-  BackstageItemUtilities,
-  UiItemsManager,
-  UiItemsProvider,
-} from "@itwin/appui-react";
-import { WidgetApiStage } from "../frontstages/WidgetApiStage";
-import { SetWidgetStateTool } from "../../tools/UiLayoutTools";
-
-export class CustomStageUiItemsProvider implements UiItemsProvider {
-  public static providerId = "appui-test-providers:custom-stage";
-  public readonly id = CustomStageUiItemsProvider.providerId;
-
-  public static register() {
-    UiItemsManager.register(new CustomStageUiItemsProvider(), {
-      stageIds: [WidgetApiStage.stageId],
-    });
-    SetWidgetStateTool.register();
-  }
-
-  public static unregister() {
-    UiItemsManager.unregister(CustomStageUiItemsProvider.providerId);
-  }
-
-  public provideBackstageItems(): BackstageItem[] {
-    return [
-      BackstageItemUtilities.createStageLauncher({
-        stageId: "example:CustomFrontstage",
-        groupPriority: 300,
-        itemPriority: 2,
-        label: "Custom Frontstage",
-      }),
-    ];
-  }
-}
