+++ test-apps/appui-test-app/standalone/src/frontend/appConfig.ts
@@ -4,8 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 function createConfig() {
   return {
-    snapshotPath: import.meta.env.IMJS_UITESTAPP_SNAPSHOT_FILEPATH,
     cesiumIonKey: import.meta.env.IMJS_CESIUM_ION_KEY,
+    snapshotPath: import.meta.env.IMJS_UITESTAPP_SNAPSHOT_FILEPATH,
+    urlPrefix: import.meta.env.IMJS_URL_PREFIX,
+    bingMapsKey: import.meta.env.IMJS_BING_MAPS_KEY,
+    mapBoxKey: import.meta.env.IMJS_MAPBOX_KEY,
   };
 }
