+++ test-apps/appui-test-app/standalone/src/frontend/appui/openBlankConnection.ts
