+++ test-apps/appui-test-app/standalone/src/frontend/appui/settingsproviders/Settings.scss
