--- test-apps/appui-test-app/standalone/src/frontend/index.scss
@@ -1,41 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-@import "~@itwin/core-react/lib/cjs/core-react/style/themecolors";
-
-body {
-  margin: 0;
-  padding: 0;
-  font-family: var(--iui-font-sans);
-  font-size: var(--iui-font-size-1);
-  color: var(--iui-color-text);
-}
-
-.red-text {
-  color: var(--iui-color-text-negative);
-}
-
-.app-portal {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-
-  &.app-overflow {
-    display: block;
-    overflow: auto;
-
-    .app-viewer {
-      height: 100vh;
-    }
-  }
-}
-
-.app-header {
-  background: #ccc;
-  height: 60px;
-}
-
-.app-viewer {
-  height: 100%;
-}
