+++ test-apps/appui-test-app/standalone/src/routes/local_.$fileName.tsx
@@ -8,13 +8,43 @@ import { PageLayout } from "@itwin/itwinui-layouts-react";
 import {
   BackstageComposer,
   ConfigurableUiContent,
+  SettingsModalFrontstage,
+  StandardContentToolsUiItemsProvider,
+  StandardNavigationToolsUiItemsProvider,
+  StandardStatusbarUiItemsProvider,
+  ThemeManager,
   UiFramework,
+  UiItemsManager,
 } from "@itwin/appui-react";
-import { createMainFrontstage } from "../frontend/appui/frontstages/MainFrontstage";
+import {
+  createMainFrontstage,
+  createMainFrontstageLauncher,
+} from "../frontend/appui/frontstages/MainFrontstage";
 import { appConfig } from "../frontend/appConfig";
 import { SnapshotConnection } from "@itwin/core-frontend";
-import { createViewState } from "../frontend/appui/frontstages/LocalFileStage";
+import { createViewState } from "../frontend/createViewState";
 import { appInitializer } from "../frontend/AppInitializer";
+import {
+  ComponentExamplesModalFrontstage,
+  FloatingWidgetsUiItemsProvider,
+} from "@itwin/appui-test-providers";
+import {
+  createElementStackingFrontstage,
+  createElementStackingProvider,
+} from "../frontend/appui/frontstages/ElementStacking";
+import { createTestPopoutFrontstage } from "../frontend/appui/frontstages/TestPopoutFrontstage";
+import { createTestPanelFrontstage } from "../frontend/appui/frontstages/TestPanelFrontstage";
+import { ProcessDetector } from "@itwin/core-bentley";
+import { EditTools } from "@itwin/editor-frontend";
+import {
+  createEditorFrontstage,
+  createEditorFrontstageProvider,
+} from "../frontend/appui/frontstages/EditorFrontstage";
+import { useEngagementTime } from "../frontend/appui/useEngagementTime";
+import {
+  createWidgetApiFrontstage,
+  createWidgetApiStageProvider,
+} from "../frontend/appui/frontstages/WidgetApiFrontstage";
 
 export const Route = createFileRoute("/local/$fileName")({
   component: Local,
@@ -24,6 +54,10 @@ export const Route = createFileRoute("/local/$fileName")({
     const filePath = `${appConfig.snapshotPath}/${ctx.params.fileName}`;
     const iModelConnection = await SnapshotConnection.openFile(filePath);
     const viewState = await createViewState(iModelConnection);
+
+    if (ProcessDetector.isElectronAppFrontend) {
+      await EditTools.initialize();
+    }
     return {
       iModelConnection,
       viewState,
@@ -40,18 +74,74 @@ function Local() {
         viewState,
       },
     });
-    const frontstages = [mainFrontstage];
+    const frontstages = [
+      mainFrontstage,
+      createElementStackingFrontstage(),
+      createTestPanelFrontstage(),
+      createTestPopoutFrontstage(),
+      createWidgetApiFrontstage(),
+    ];
     frontstages.forEach((frontstage) => {
-      void UiFramework.frontstages.addFrontstage(frontstage);
+      UiFramework.frontstages.addFrontstage(frontstage);
     });
-    void UiFramework.frontstages.setActiveFrontstage(mainFrontstage.id);
+
+    UiItemsManager.register(new StandardContentToolsUiItemsProvider(), {
+      stageIds: [
+        createMainFrontstage.stageId,
+        createEditorFrontstage.stageId,
+        createWidgetApiFrontstage.stageId,
+      ],
+    });
+    UiItemsManager.register(new StandardNavigationToolsUiItemsProvider(), {
+      stageIds: [
+        createMainFrontstage.stageId,
+        createWidgetApiFrontstage.stageId,
+      ],
+    });
+    UiItemsManager.register(new StandardStatusbarUiItemsProvider(), {
+      stageIds: [
+        createMainFrontstage.stageId,
+        createWidgetApiFrontstage.stageId,
+      ],
+    });
+    UiItemsManager.register({
+      id: "appui-test-app:backstageItemsProvider",
+      getBackstageItems: () => [
+        createMainFrontstageLauncher(),
+        SettingsModalFrontstage.getBackstageActionItem(400, 10),
+        ComponentExamplesModalFrontstage.getBackstageActionItem(400, 20),
+      ],
+    });
+    UiItemsManager.register(createElementStackingProvider(), {
+      stageIds: [createElementStackingFrontstage.stageId],
+    });
+    UiItemsManager.register(new FloatingWidgetsUiItemsProvider(), {
+      stageIds: [createWidgetApiFrontstage.stageId],
+    });
+    UiItemsManager.register(createWidgetApiStageProvider(), {
+      stageIds: [createWidgetApiFrontstage.stageId],
+    });
+
+    if (ProcessDetector.isElectronAppFrontend) {
+      UiFramework.frontstages.addFrontstage(createEditorFrontstage());
+      UiItemsManager.register(createEditorFrontstageProvider(), {
+        stageIds: [createEditorFrontstage.stageId],
+      });
+    }
+
+    void UiFramework.frontstages.setActiveFrontstage(
+      createMainFrontstage.stageId
+    );
     return () => {
       UiFramework.frontstages.clearFrontstageDefs();
     };
   }, [iModelConnection, viewState]);
+  useEngagementTime();
   return (
     <PageLayout.Content>
-      <ConfigurableUiContent appBackstage={<BackstageComposer />} />
+      <ThemeManager>
+        <ConfigurableUiContent appBackstage={<BackstageComposer />} />
+      </ThemeManager>
     </PageLayout.Content>
   );
 }
